Backend_backup/requirements.txt
beautifulsoup4
pillow
pytesseract
datasketch
flask
pdf2image
flask-cors
